+++ app/src/main/java/com/victor_sml/playlistmaker/presentation/IterativeLambda.kt
@@ -1,6 +1,9 @@
 package com.victor_sml.playlistmaker.presentation
 
 interface IterativeLambda {
-    fun start(delayMillis: Long, lambda: () -> Unit)
+    val delayMillis: Long
+    val lambda: () -> Unit
+
+    fun start()
     fun stop()
 }
